+++ esp_video/private_include/esp_video_device_internal.h
@@ -22,7 +22,6 @@
 extern "C" {
 #endif
 
-
 /**
  * @brief MIPI-CSI state
  */