+++ esp_video/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add the initialization parameter "dont_init_ldo" to disable LDO initialization for the MIPI-CSI video device
 - Add WBG(white balance gain) ioctl command to ISP video device
+- Add BLC(Black level correction) ioctl command to ISP video device
 
 ## 1.3.1